update without xtokens
